+++ test/clojure_lsp/api_test.clj
@@ -21,7 +21,13 @@
       (reset! db/db {})
       (with-redefs [spit #(is (= (slurp "integration-test/sample-test/fixtures/api/clean_ns/a.clj") %2))]
         (api/clean-ns! :project-root (io/file "integration-test/sample-test")
-                       :namespace '[api.clean-ns.a])))))
+                       :namespace '[api.clean-ns.a])))
+    (testing "when a single namespace is specified with dry option"
+      (reset! db/db {})
+      (is (thrown? clojure.lang.ExceptionInfo
+            (api/clean-ns! :project-root (io/file "integration-test/sample-test")
+                           :namespace '[api.clean-ns.a]
+                           :dry? true))))))
 
 (deftest rename!
   (testing "when project-root is not a file"