+++ test/clojure_lsp/main_test.clj
@@ -17,6 +17,9 @@
     (testing "log-path"
       (is (= "/custom/path" (:log-path (:options (#'main/parse ["--log-path" "/custom/path"])))))
       (is (= nil (:log-path (:options (#'main/parse []))))))
+    (testing "dry"
+      (is (not (:dry? (:options (#'main/parse [])))))
+      (is (:dry? (:options (#'main/parse ["--dry"])))))
     (testing "project-root"
       (is (= default-root (.getAbsolutePath (:project-root (:options (#'main/parse ["--project-root" "src"]))))))
       (is (= default-root (.getAbsolutePath (:project-root (:options (#'main/parse ["-p" "src"]))))))